renaming test case classes
