+++ src/main/java/com/linuxforhealth/connect/builder/BlueButton20RestRouteBuilder.java
@@ -11,7 +11,13 @@ import com.linuxforhealth.connect.processor.BlueButton20MetadataProcessor;
 import com.linuxforhealth.connect.processor.BlueButton20RequestProcessor;
 import com.linuxforhealth.connect.processor.BlueButton20ResultProcessor;
 import com.linuxforhealth.connect.support.CamelContextSupport;
+import org.apache.camel.Exchange;
+import org.apache.camel.Processor;
 import org.apache.camel.builder.RouteBuilder;
+import org.apache.camel.builder.SimpleBuilder;
+import org.apache.commons.lang3.SystemUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.net.URI;
 
@@ -23,6 +29,9 @@ public class BlueButton20RestRouteBuilder extends RouteBuilder {
     public final static String AUTHORIZE_ROUTE_ID = "bluebutton-20-rest-authorize";
     public final static String CALLBACK_ROUTE_ID = "bluebutton-20-rest-callback";
     public final static String API_ROUTE_ID = "bluebutton-20-rest";
+
+    private final Logger logger = LoggerFactory.getLogger(BlueButton20AuthProcessor.class);
+
 
     @Override
     public void configure() {
@@ -40,13 +49,40 @@ public class BlueButton20RestRouteBuilder extends RouteBuilder {
                 .get()
                 .route()
                 .routeId(AUTHORIZE_ROUTE_ID)
-                .doTry()
-                    .process( new BlueButton20AuthProcessor())
-                    .toD("${exchangeProperty[location]}")
-                .doCatch(Exception.class)
-                    .setProperty("errorMessage", simple(exceptionMessage().toString()))
-                    .to(LinuxForHealthRouteBuilder.ERROR_CONSUMER_URI)
-                .end();
+                .process(exchange -> {
+                    String callbackURL = SimpleBuilder
+                            .simple("{{lfh.connect.bluebutton_20.handlerUri}}")
+                            .evaluate(exchange, String.class);
+
+                    String cmsAuthorizeURL = SimpleBuilder
+                            .simple("{{lfh.connect.bluebutton_20.cms.authorizeUri}}")
+                            .evaluate(exchange, String.class);
+
+                    String clientId = SimpleBuilder
+                            .simple("{{lfh.connect.bluebutton_20.cms.clientId}}")
+                            .evaluate(exchange, String.class);
+
+                    // Set up call to redirect to Blue Button API so the user can authenticate this application
+                    String authorizeURL = cmsAuthorizeURL +
+                            "?client_id=" + clientId+
+                            "&redirect_uri=" + callbackURL +
+                            "&response_type=code";
+
+                    logger.debug("Authorize URL: "+authorizeURL);
+
+                    // Determine the current OS so we know the cmd to launch the browser
+                    String osCmd;
+                    if (SystemUtils.IS_OS_MAC) {
+                        osCmd = "open";
+                    } else if (SystemUtils.IS_OS_WINDOWS) {
+                        osCmd = "explorer";
+                    } else {
+                        // Assume SystemUtils.IS_OS_UNIX
+                        osCmd = "xdg-open";
+                    }
+                    exchange.setProperty("location", "exec:"+osCmd+"?args=RAW("+authorizeURL+")");
+                })
+                .toD("${exchangeProperty[location]}");
 
         // Blue Button OAuth2 - Callback to exchange code for token (displayed in the browser)
         URI blueButtonHandlerUri = URI.create(contextSupport.getProperty("lfh.connect.bluebutton_20.handlerUri"));
