+++ src/test/java/com/linuxforhealth/connect/builder/BlueButton20AuthTest.java
@@ -1,0 +1,4 @@
+package com.linuxforhealth.connect.builder;
+
+public class BlueButton20AuthTest {
+}
