+++ src/test/java/com/linuxforhealth/connect/builder/LinuxForHealthErrorTest.java
@@ -7,23 +7,20 @@ package com.linuxforhealth.connect.builder;
 
 import com.linuxforhealth.connect.support.TestUtils;
 import org.apache.camel.RoutesBuilder;
-import org.apache.camel.component.kafka.KafkaConstants;
 import org.apache.camel.component.mock.MockEndpoint;
-import org.apache.kafka.clients.producer.RecordMetadata;
-import org.apache.kafka.common.TopicPartition;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Properties;
 
 /**
- * Tests {@link LinuxForHealthRouteBuilder#NOTIFY_CONSUMER_URI}
+ * Tests {@link LinuxForHealthRouteBuilder#ERROR_CONSUMER_URI}
  */
-public class LinuxForHealthNotifyRouteBuilderTest extends RouteBuilderTestSupport {
+public class LinuxForHealthErrorTest extends RouteTestSupport {
 
+    private MockEndpoint mockDataStoreResult;
     private MockEndpoint mockMessagingResult;
+    private MockEndpoint mockUnreachableResult;
 
     /**
      * Provides properties to support mocking data and messaging components.
@@ -33,10 +30,11 @@ public class LinuxForHealthNotifyRouteBuilderTest extends RouteBuilderTestSuppor
     protected Properties useOverridePropertiesWithPropertiesComponent() {
         Properties props = super.useOverridePropertiesWithPropertiesComponent();
 
-        props.setProperty("lfh.connect.test.uri", "direct:test-notify");
+        props.setProperty("lfh.connect.test.uri", "direct:test-error");
         props.setProperty("lfh.connect.test.dataFormat", "csv");
         props.setProperty("lfh.connect.test.messageType", "person");
 
+        props.setProperty("lfh.connect.dataStore.uri", "mock:data-store");
         props.setProperty("lfh.connect.messaging.uri", "mock:messaging");
         return props;
     }
@@ -51,26 +49,17 @@ public class LinuxForHealthNotifyRouteBuilderTest extends RouteBuilderTestSuppor
                 new LinuxForHealthRouteBuilder(),
                 new BaseRouteBuilder() {
                     @Override
-                    protected String getRoutePropertyNamespace() { return "lfh.connect.test"; }
+                    protected String getRoutePropertyNamespace() {
+                        return "lfh.connect.test";
+                    }
+
                     @Override
                     protected void buildRoute(String routePropertyNamespace) {
+                        // define a route which will throw an exception, triggering the error route
                         from("{{lfh.connect.test.uri}}")
-                        .routeId("test-notify")
-                        .process(exchange -> {
-                            exchange.setProperty("uuid", "312e02e6-9a33-494a-8928-1aba9dbedf9c");
-                            exchange.setProperty("routeUri", "http://0.0.0.0:8080/fhir/r4/PATIENT");
-                            exchange.setProperty("dataFormat", "fhir-r4");
-                            exchange.setProperty("timestamp", 1596830090);
-                            exchange.setProperty("dataStoreUri", "kafka:FHIR_R4_PATIENT?brokers=localhost:9092");
-                            exchange.setProperty("status", "success");
-
-                            RecordMetadata rm =  new RecordMetadata(
-                                    new TopicPartition("FHIR-R4_PATIENT", 0), 0, 0, 1591732928186L, 0L, 0, 0);
-                            List<RecordMetadata> metaRecords = new ArrayList<>();
-                            metaRecords.add(rm);
-                            exchange.getIn().setHeader(KafkaConstants.KAFKA_RECORDMETA, metaRecords);
-                        })
-                        .to(LinuxForHealthRouteBuilder.NOTIFY_CONSUMER_URI);
+                        .routeId("test-error-handler")
+                        .throwException(RuntimeException.class, "runtime exception")
+                        .to("mock:test-error");
                     }
                 }
         };
@@ -83,18 +72,31 @@ public class LinuxForHealthNotifyRouteBuilderTest extends RouteBuilderTestSuppor
     @Override
     protected void configureContext() throws Exception {
         super.configureContext();
+        mockDataStoreResult = MockEndpoint.resolve(context, "mock:data-store");
         mockMessagingResult = MockEndpoint.resolve(context, "mock:messaging");
+        mockUnreachableResult = MockEndpoint.resolve(context, "mock:test-error");
     }
 
+    /**
+     * Tests {@link LinuxForHealthRouteBuilder#ERROR_CONSUMER_URI}
+      * @throws Exception
+     */
     @Test
-    void testNotifyRoute() throws Exception {
+    void testErrorRoute() throws Exception {
+        mockUnreachableResult.expectedMessageCount(0);
+        mockDataStoreResult.expectedMessageCount(0);
+
         String expectedMsg = context
                 .getTypeConverter()
-                .convertTo(String.class, TestUtils.getMessage("lfh", "notify.json"));
-        mockMessagingResult.expectedMessageCount(1);
-        mockMessagingResult.expectedBodiesReceived(expectedMsg);
+                .convertTo(String.class, TestUtils.getMessage("lfh", "error.json"));
 
-        producerTemplate.sendBody("direct:test-notify", "notify test message");
+        mockMessagingResult.expectedBodiesReceived(expectedMsg);
+        mockMessagingResult.expectedMessageCount(1);
+
+        producerTemplate.sendBody("direct:test-error", "1,Donald,Duck");
+
+        mockUnreachableResult.assertIsSatisfied();
+        mockDataStoreResult.assertIsSatisfied();
         mockMessagingResult.assertIsSatisfied();
-    }
+   }
 }
